+++ redbot/core/commands/help.py
@@ -39,8 +39,7 @@ from discord.ext import commands as dpy_commands
 from . import commands
 from .context import Context
 from ..i18n import Translator
-from ..utils.help import HelpSource
-from ..utils._dpy_menus_utils import SimpleHybridMenu
+from ..utils._dpy_menus_utils import SimpleHybridMenu, HelpSource
 from ..utils.mod import mass_purge
 from ..utils._internal_utils import fuzzy_command_search, format_fuzzy_results
 from ..utils.chat_formatting import box, humanize_list, humanize_number, humanize_timedelta, pagify
